refactor: android base map view refactoring
