+++ android/src/main/kotlin/com/google/maps/flutter/navigation/GoogleMapsBaseMapView.kt
@@ -25,6 +25,7 @@ import com.google.android.gms.maps.CameraUpdate
 import com.google.android.gms.maps.CameraUpdateFactory
 import com.google.android.gms.maps.GoogleMap
 import com.google.android.gms.maps.GoogleMap.OnMarkerDragListener
+import com.google.android.gms.maps.GoogleMapOptions
 import com.google.android.gms.maps.model.CameraPosition
 import com.google.android.gms.maps.model.Circle
 import com.google.android.gms.maps.model.FollowMyLocationOptions
@@ -38,6 +39,7 @@ import com.google.android.libraries.navigation.NavigationView
 
 abstract class GoogleMapsBaseMapView(
   protected val viewId: Int,
+  mapOptions: GoogleMapOptions,
   protected val viewEventApi: ViewEventApi,
   private val imageRegistry: ImageRegistry
 ) {
@@ -51,6 +53,11 @@ abstract class GoogleMapsBaseMapView(
   private val _polygons = mutableListOf<PolygonController>()
   private val _polylines = mutableListOf<PolylineController>()
   private val _circles = mutableListOf<CircleController>()
+
+  // Store preferred zoom values here because MapView getMinZoom and
+  // getMaxZoom always return min/max possible values and not the preferred ones.
+  private var _minZoomLevelPreference: Float? = null
+  private var _maxZoomLevelPreference: Float? = null
 
   // Nullable variable to hold the callback function
   private var _mapReadyCallback: ((Result<Unit>) -> Unit)? = null
@@ -81,6 +88,11 @@ abstract class GoogleMapsBaseMapView(
     } else {
       throw FlutterError("mapNotFound", "GoogleMap not initialized yet")
     }
+  }
+
+  init {
+    _minZoomLevelPreference = mapOptions.minZoomPreference
+    _maxZoomLevelPreference = mapOptions.maxZoomPreference
   }
 
   protected fun mapReady() {
@@ -299,6 +311,46 @@ abstract class GoogleMapsBaseMapView(
   fun setZoomControlsEnabled(enabled: Boolean) {
     invalidateViewAfterMapLoad()
     getMap().uiSettings.isZoomControlsEnabled = enabled
+  }
+
+  fun getMinZoomPreference(): Float {
+    return _minZoomLevelPreference ?: getMap().minZoomLevel
+  }
+
+  fun getMaxZoomPreference(): Float {
+    return _maxZoomLevelPreference ?: getMap().maxZoomLevel
+  }
+
+  fun resetMinMaxZoomPreference() {
+    _minZoomLevelPreference = null
+    _maxZoomLevelPreference = null
+    getMap().resetMinMaxZoomPreference()
+  }
+
+  @Throws(FlutterError::class)
+  fun setMinZoomPreference(minZoomPreference: Float) {
+    if (minZoomPreference > (_maxZoomLevelPreference ?: getMap().maxZoomLevel)) {
+      throw FlutterError(
+        "minZoomGreaterThanMaxZoom",
+        "Minimum zoom level cannot be greater than maximum zoom level"
+      )
+    }
+
+    _minZoomLevelPreference = minZoomPreference
+    getMap().setMinZoomPreference(minZoomPreference)
+  }
+
+  @Throws(FlutterError::class)
+  fun setMaxZoomPreference(maxZoomPreference: Float) {
+    if (maxZoomPreference < (_minZoomLevelPreference ?: getMap().minZoomLevel)) {
+      throw FlutterError(
+        "maxZoomLessThanMinZoom",
+        "Maximum zoom level cannot be less than minimum zoom level"
+      )
+    }
+
+    _maxZoomLevelPreference = maxZoomPreference
+    getMap().setMaxZoomPreference(maxZoomPreference)
   }
 
   fun setCompassEnabled(enabled: Boolean) {
